Working with new format of outline extraction
